clustercat.{c,h}: use doubles for global-level scores

+++ src/clustercat.h
@@ -69,7 +69,7 @@ void tokenize_sent(char * restrict sent_str, struct_sent_info *sent_info, bool c
 void init_clusters(const struct cmd_args cmd_args, unsigned long vocab_size, char **unique_words, struct_map_word_class **word2class_map);
 void cluster(const struct cmd_args cmd_args, char * restrict sent_store[const], const struct_model_metadata model_metadata, char **unique_words, struct_map **word_map, struct_map_word_class **word2class_map);
 struct_sent_info parse_input_line(char * restrict line_in, struct_map **ngram_map);
-float query_sents_in_store(const struct cmd_args cmd_args, char * restrict sent_store[const], const struct_model_metadata model_metadata, struct_map **ngram_map, struct_map_class **class_map, struct_map_word_class **word2class_map);
+double query_sents_in_store(const struct cmd_args cmd_args, char * restrict sent_store[const], const struct_model_metadata model_metadata, struct_map **ngram_map, struct_map_class **class_map, struct_map_word_class **word2class_map);
 
 void print_sent_info(struct_sent_info * restrict sent_info);
 #endif // INCLUDE_HEADER
